# app/swagger.yml
swagger: "2.0"
info:
  description: This is the swagger file that goes with our server code, all api start with "url:5000/api/<name>", for example "http://localhost:5000/api/user"
  version: "1.0.0"
  title: Swagger ReST Article
consumes:
  - application/json
produces:
  - application/json
basePath: /api
...
paths:
  /user:
    get:
      operationId: api.user.read_list
      tags:
        - User
      summary: Read list of users
      description: Read list of N users if user that call this API is not logged, otherwise return the first N users giving importance to users near the user that call this API. For this reason we pass to this API also the id of user that call it
      parameters:
      - name: token
        in: header
        description: Token generated for user during login
        required: False
        type: string
      - name: id
        in: header
        description: Id that identify user
        required: False
        type: integer
      - name: language
        in: query
        description: Language using in client
        required: True
        type: string
      responses:
        200:
          description: Successfully read people list operation
          schema:
            type: object
            properties:
              status: 
                type: integer
                description: status returned by api
              users:
                type: array
                description: list of N users with basic information show in travelers page
                items:
                  type: object
                  properties:
                    id: 
                      type: integer
                      description: Id of user
                    username: 
                      type: string
                      description: username of user
                    image: 
                      type: string
                      description: path of profile image of user
                    languages:
                      type: array
                      items:
                        type: object
                        properties:
                          language:
                            type: string
                            description: name of the language
                    home:
                      type: object
                      properties:
                        country:
                          type: string
                          description: country of locality visited by user
                        provice:
                          type: string
                          description: provice of locality visited by user
                        city:
                          type: string
                          description: city of locality visited by user
                
    post:
      operationId: api.user.create
      tags:
        - User
      summary: Create a new user
      description: Create a new user with basic information
      parameters:
        - name: language
          in: query
          description: Language using in client
          required: True
          type: string
        - name: data
          in: body
          description: all basic datas for the creation of a new user
          required: True
          schema:
            type: object
            properties:
              email:
                type: string
              username:
                type: string
              password:
                type: string
      responses:
        201:
          description: Successfully created new user
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
  
    delete:
      operationId: api.user.delete
      tags:
        - User
      summary: Delete 1 user
      description: Delete the user with the id passed
      parameters:
        - name: language
          in: query
          description: Language using in client
          required: True
          type: string
        - name: token
          in: header
          type: string
          description: Token created when user logged, send to verify if it is valid
          required: True
      responses:
        200: 
          description: Successfully deleted person from user table
          schema:
            type: object
            properties:
              message: 
                type: string
                description: message sent by server
              status: 
                type: integer
                description: status returned by api
...
  /tip/list:
    post:
      operationId: api.tip.read_list
      tags:
        - Tip
      summary: Get list of tips
      description: Get list of tips appling filters
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: False
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: data
          in: body
          description: Filters passed by the client to get only the desired tips
          required: False
          schema:
            type: object
            properties:
              own:
                type: boolean
                description: true if user want to se only own tips
              username:
                type: string
              country:
                type: string
              province:
                type: string
              city:
                type: string
              tags:
                type: string
                description: list of tag separated by comma
              dates:
                type: string
              text:
                type: string
                description: look for a word in title or text
              languages:
                type: string
                description: look for tips written in the same language/languages
      responses:
        200:
          description: Tips list successfully get
...
  /tip/vote:
    put:
      operationId: api.tip.update_vote
      tags:
        - Tip
      summary: Update 1 vote
      description: Update 1 vote user does for 1 tip
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: idTip
          in: query
          description: Id that identify the tip voted
          type: integer
          required: True
        - name: vote
          in: query
          description: Vote given (from 0 to 4)
          type: integer
          required: False
      responses:
          200:
            description: Vote for tip updated successfully
  /itinerary/list:
    post:
      operationId: api.itinerary.read_list
      tags:
        - Itinerary
      summary: Get list of itineraries
      description: Get list of itineraries appling filters
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: False
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: data
          in: body
          description: Filters passed by the client to get only the desired itineraries
          required: False
          schema:
            type: object
            properties:
              own:
                type: boolean
                description: true if user want to se only own itineraries
              username:
                type: string
              country:
                type: string
              province:
                type: string
              city:
                type: string
              tags:
                type: string
                description: list of tag separated by comma
              dates:
                type: string
              text:
                type: string
                description: look for a word in title or text
              languages:
                type: string
                description: look for itineraries written in the same language/languages
      responses:
        200:
          description: Tips list successfully get
...
  /itinerary/vote:
    put:
      operationId: api.itinerary.update_vote
      tags:
        - Itinerary
      summary: Update 1 vote
      description: Update 1 vote user does for 1 itinerary
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: idItinerary
          in: query
          description: Id that identify the itinerary voted
          type: integer
          required: True
        - name: vote
          in: query
          description: Vote given (from 0 to 4)
          type: integer
          required: False
      responses:
          200:
            description: Vote for itinerary updated successfully 
  /experience/list:
    post:
      operationId: api.experience.read_list
      tags:
        - Experience
      summary: Get list of experiences
      description: Get list of experiences appling filters
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: False
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: data
          in: body
          description: Filters passed by the client to get only the desired experiences
          required: False
          schema:
            type: object
            properties:
              own:
                type: boolean
                description: true if user want to se only own experiences
              username:
                type: string
              country:
                type: string
              province:
                type: string
              city:
                type: string
              tags:
                type: string
                description: list of tag separated by comma
              dates:
                type: string
              text:
                type: string
                description: look for a word in title or text
              languages:
                type: string
                description: look for experiences written in the same language/languages
      responses:
        200:
          description: Tips list successfully get
  /experience/listBooked:
    post:
      operationId: api.experience.read_list_booked
      tags:
        - Experience
      summary: Get list of experiences
      description: Get list of experiences appling filters
      parameters:
        - name: token
          in: header
          type: string
          description: Token created when user logged
          required: True
        - name: language
          in: query
          description: Language using in client
          type: string
          required: True
        - name: id
          in: header
          type: integer
          description: User Id connected to the token
          required: True
        - name: data
          in: body
          description: Filters passed by the client to get only the desired experiences
          required: False
          schema:
            type: object
            properties:
              own:
                type: boolean
                description: true if user want to se only own experiences
              username:
                type: string
              country:
                type: string
              province:
                type: string
              city:
                type: string
              tags:
                type: string
                description: list of tag separated by comma
              dates:
                type: string
              text:
                type: string
                description: look for a word in title or text
              languages:
                type: string
                description: look for experiences written in the same language/languages
      responses:
        200:
          description: Tips list successfully get
...